Correct test to use the right class (ConfigManager.configmanager) and adjust issues with TinyDB not closing itself.

+++ newtest.py
@@ -2,7 +2,7 @@ import unittest
 import os
 import pathlib
 
-import settingsdb
+import configmanager
 
 TESTDIR = pathlib.Path(os.path.curdir).resolve()
 TESTFILE = TESTDIR / "db.json"
@@ -11,45 +11,40 @@ class FirstTest(unittest.TestCase):
     c = None
 
     def setUp(self):
-        os.remove(str(TESTFILE))
+        if os.path.exists(str(TESTFILE)):
+            os.remove(str(TESTFILE))
 
     def test1(self):
         # testing blank initilization
-        c = settingsdb.setting()
+        c = configmanager.ConfigManager()
         c.initialize(TESTDIR)
+        c.db.close()
 
         self.assertTrue(c.setup)
 
-        c.db.close()
 
     def test2(self):
         # test initilization and then create with same path.
 
-        c = settingsdb.setting()
+        c = configmanager.ConfigManager()
         c.initialize(TESTDIR)
-        self.assertTrue(c.setup)
         c.db.close()
-
-        c = settingsdb.setting(TESTDIR)
         self.assertTrue(c.setup)
 
+        c = configmanager.ConfigManager(TESTDIR)
         c.db.close()
-
+        self.assertTrue(c.setup)
 
     def test3(self):
-        c = settingsdb.setting()
+        # test initializing it, and making sure the test file is where we expect it to be
+
+        c = configmanager.ConfigManager()
         c.initialize(TESTDIR)
         c.db.close()
         self.assertTrue(c.setup)
-
-        self.assertEqual(c.dbfile,TESTFILE)
+        self.assertEqual(c.dbpath,TESTDIR)
         
-        c.db.close()
-        
-
     
-
-
 
 if __name__ == "__main__":
     unittest.main()
